// client/src/game/shared.ts
export const MAPS = ['m1', 'm2'] as const;
export type MapLocation = typeof MAPS[number];

export interface ILocation {
  pos: { x: number; y: number };
  map: MapLocation;
}

export interface ICreateAccount {
  username: string;
  password: string;
  dna: IUserDNA;
}

export type UserSprite = 'c1';
// ...
export type AttackType = 'slash' | 'thrust' | 'spell' | 'oversize' | 'shoot';

export interface IEnemyAttacked {
  id: string;
  dead: boolean;
  damage: number;
  attributes: StrictMap<CombatAttribute, number>;
  loot: IItem[];
}

export interface IUserAttack {
  name: string;
  enemies: IEnemyAttacked[];
}

export interface IUpdateLoot {
  id: string;
  items: IItem[];
}

export const COMBAT_ATTRIBUTES = ['health', 'stamina', 'mana'] as const;
export type CombatAttribute = typeof COMBAT_ATTRIBUTES[number];

export const LIFE_ATTRIBUTES = ['hunger', 'exhaustion', 'age'] as const;
export type LifeAttribute = typeof LIFE_ATTRIBUTES[number];

export type UserAttribute = LifeAttribute | CombatAttribute;

export const COMBAT_STATS = [
  'health',
  'stamina',
  'mana',
  'agility',
  'defense',
  'accuracy',
  'strength',
  'intelligence',
  'dexterity',
  'sneak',
  'awareness',
  'slash',
  'thrust',
  'spell',
  'oversize',
  'shoot',
] as const;
export type CombatStat = typeof COMBAT_STATS[number];

export const GATHERING_STATS = [
  'woodcutting',
  'mining',
  'fishing',
  'farming',
] as const;
export type GatheringStat = typeof GATHERING_STATS[number];

export const ARTISAN_STATS = [
  'cooking',
  'crafting',
  'enchanting',
  'alchemy',
] as const;
export type ArtisanStat = typeof ARTISAN_STATS[number];

export const SUPPORT_STATS = ['thieving'] as const;
export type SupportStat = typeof SUPPORT_STATS[number];

export const USER_STATS = [
  ...COMBAT_STATS,
  ...GATHERING_STATS,
  ...ARTISAN_STATS,
  ...SUPPORT_STATS,
] as const;

export type UserStat = typeof USER_STATS[number];

export interface IStat {
  level: number;
  xp: number;
  nextXp: number;
  points: number;
}

const ITEM_TYPES = ['WEARABLE', 'USEABLE', 'CONSUMABLE'] as const;
type ItemType = typeof ITEM_TYPES[number];

export interface IEquipItem {
  name: string;
  equipement: Partial<IUserEquipement>;
  inventory?: IItem[];
}

export interface IItem {
  name: string;
  sprite: string;
  quantity: number;
  type: ItemType;
  sex?: 'male' | 'female';
  bonus?: Partial<StrictMap<UserStat, number>>;
  duration?: number;
  activateJob?: GatheringStat | ArtisanStat;
  equipSlot?: EquipementSlot;
  equipSprite?: string;
  attackType?: AttackType;
}

export const EQUIPEMENT_SLOTS = [
  'earring1',
  'head',
  'earring2',
  'shoulders',
  'torso',
  'cape',
  'necklace',
  'weapon',
  'arms',
  'belt',
  'hands',
  'shield',
  'ring1',
  'legs',
  'feet',
  'ring2',
] as const;

export type EquipementSlot = typeof EQUIPEMENT_SLOTS[number];

export class IUserEquipement implements StrictMap<EquipementSlot, any> {
  feet: IItem;
  legs: IItem;
  belt: IItem;
  torso: IItem;
  shoulders: IItem;
  arms: IItem;
  hands: IItem;
  cape: IItem;
  head: IItem;
  weapon: IItem;
  shield: IItem;
  ring1: IItem;
  ring2: IItem;
  necklace: IItem;
  earring1: IItem;
  earring2: IItem;
}

export interface IUserDNA {
  sex: 'female' | 'male';
  skin:
    | 'white'
    | 'black'
    | 'olive'
    | 'brown'
    | 'peach'
    | 'dark'
    | 'dark2'
    | 'tan1'
    | 'tan2';

  eyes:
    | 'blue'
    | 'brown'
    | 'gray'
    | 'green'
    | 'purple'
    | 'red'
    | 'yellow'
    | 'orange';

  facialHair: string;
  hair: string;
  hairColor: string;
  nose: string;
  ears: string;
}

export interface IUser {
  dna: IUserDNA;
  location: ILocation;
  sprite: UserSprite;
  name: string;
  stats: StrictMap<UserStat, IStat>;
  attributes: StrictMap<UserAttribute, number>;
  threat: number;
  inventory: IItem[];
  equipement: Partial<IUserEquipement>;
}